AI_Agent/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
google-generativeai
chromadb
openai-whisper
sounddevice
numpy
soundfile
python-dotenv
pydantic